Cleanup code, add scripts for save and load game functionality
