+++ Assets/Scenes/GameScene.unity
@@ -1211,16 +1211,28 @@ MonoBehaviour:
   playerCamera: {fileID: 963194227}
   interactionRange: 7
   interactionText: {fileID: 369568432}
-  mainMenu: {fileID: 392281866}
-  infoMenu: {fileID: 1906331903}
-  quizMenu: {fileID: 1761021858}
   exhibitLayer:
     serializedVersion: 2
     m_Bits: 64
+  MainMenu: {fileID: 392281866}
+  infoMenu: {fileID: 1906331903}
+  quizMenu: {fileID: 1761021858}
+  pauseMenu: {fileID: 0}
   exhibitMainMenu: {fileID: 392281870}
   exhibitInfoMenu: {fileID: 1906331907}
   exhibitQuizMenu: {fileID: 1761021862}
-  AchievementsMenu: {fileID: 1807608840}
+--- !u!114 &151721233
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 151721227}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c2aeb8dac127da14ba3d919784657526, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   playerName: {fileID: 1295514226}
 --- !u!1 &159434125
 GameObject:
@@ -2944,7 +2956,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &457533364
 RectTransform:
   m_ObjectHideFlags: 0
@@ -6632,7 +6644,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 927783754}
   serializedVersion: 2
-  m_LocalRotation: {x: 0.14521313, y: 0.00000003296683, z: -0.000000004838502, w: 0.98940045}
+  m_LocalRotation: {x: 0.14521313, y: 0.000000032966827, z: -0.0000000048385016, w: 0.98940045}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -8560,6 +8572,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: Player
       objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 25cc4da2c099df442bc519515c5807cd, type: 3}
+      propertyPath: m_TagString
+      value: Player
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects:
@@ -8579,6 +8595,9 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 25cc4da2c099df442bc519515c5807cd, type: 3}
       insertIndex: -1
       addedObject: {fileID: 151721232}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 25cc4da2c099df442bc519515c5807cd, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 151721233}
   m_SourcePrefab: {fileID: 100100000, guid: 25cc4da2c099df442bc519515c5807cd, type: 3}
 --- !u!1 &1264793414
 GameObject:
@@ -8988,7 +9007,7 @@ GameObject:
   - component: {fileID: 1295514226}
   - component: {fileID: 1295514225}
   m_Layer: 5
-  m_Name: PlayerName
+  m_Name: PlayerNameText
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -12612,7 +12631,7 @@ GameObject:
   - component: {fileID: 1914458524}
   - component: {fileID: 1914458523}
   m_Layer: 5
-  m_Name: BakToMainMenuBtn
+  m_Name: BackToMainMenuBtn
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
