+++ Assets/Scripts/PlayerScripts/PlayerInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2aeb8dac127da14ba3d919784657526
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
