+++ Assets/Scripts/PlayerScripts/SaveData.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SaveData : MonoBehaviour
+{
+    public float[] playerPosition;
+    public string playerName;
+    public List<string> unlockedAchievements;
+}
