+++ Assets/Scripts/PlayerScripts/SaveData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b2647336a28ac74b80baee06cd84d4a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
