+++ Assets/Scripts/PlayerScripts/SaveManager.cs
@@ -1,0 +1,8 @@
+﻿using System.IO;
+using UnityEngine;
+using System.Runtime.Serialization.Formatters.Binary;
+
+public static class SaveManager
+{
+    
+}
