+++ Assets/Scripts/PlayerScripts/SaveManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2411b948daabc34abb7adb732500704
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
