+++ Assets/Scripts/UIScripts/PauseMenuScript.cs
@@ -9,8 +9,6 @@ public class PauseMenuScript : MonoBehaviour
     public GameObject PauseMenu;
     public List<GameObject> otherMenus;
 
-    private int slot = 0;
-
     private void Update()
     {
         HandleInput();
